Moved body to a different component.
