+++ FrontEnd/src/Pages/ControlPanel/PagesCP/Alumnos/Alumnos.jsx
@@ -2,29 +2,20 @@ import React, { useState, useEffect, useMemo } from 'react';
 import {
   Button, Modal, TextField, Box, Typography,
 } from '@mui/material';
-import { grey } from '@mui/material/colors';
-import { DataGrid, gridClasses } from '@mui/x-data-grid';
-import Card from '@mui/material/Card';
-import CardContent from '@mui/material/CardContent';
-import { CSVLink } from 'react-csv';
-import Select from 'react-select';
-import Actions from '../../../../Components/ControlPanel/Alumnos/ActAlumnos';
 import { deleteStudent, getStudents, updateStudent } from '../../../../api/students';
 import { getPeriodos } from '../../../../api/Periodos';
 import { getClassStudent } from '../../../../api/classStudent';
 import { alumnoVacio } from '../../../../utils/constants';
 import MasInformacionAlumno from '../../../../Components/ControlPanel/Alumnos/MasInformacionAlumno';
 import HeaderAlumnos from '../../../../Components/ControlPanel/Alumnos/HeaderAlumnos';
+import BodyAlumnos from '../../../../Components/ControlPanel/Alumnos/BodyAlumnos';
 
 export default function Alumnos() {
   const [studentData, setStudentData] = useState([]);
   const [originalStudentData, setOriginalStudentData] = useState([]);
   const [dataPeriodo, setDataPeriodo] = useState([]);
   const [dataAlumnoClase, setDataAlumnoClase] = useState([]);
-  const [items, setItems] = useState([]);
   const [alumnoSeleccionado, setAlumnoSeleccionado] = useState(alumnoVacio);
-  const [pageSize, SetPageSize] = useState(5);
-
   const [openModal, setOpenModal] = useState(false);
   const [currentOperation, setCurrentOperation] = useState('');
 
@@ -68,7 +59,7 @@ export default function Alumnos() {
     setOpenModal(!openModal);
   };
 
-  const seleccionarConsola = (consola, caso) => {
+  const seleccionarAlumno = (consola, caso) => {
     setAlumnoSeleccionado(consola);
     setCurrentOperation(caso);
     abrirCerrarModal();
@@ -240,96 +231,6 @@ export default function Alumnos() {
     </div>
   );
 
-  const columns = useMemo(
-    () => [
-      {
-        field: '_id', headerName: 'Id', width: 54, hide: true,
-      },
-      {
-        field: 'clave_unica_identificacion',
-        headerName: 'clave_unica_identificacion',
-        width: 54,
-        hide: true,
-      },
-      {
-        field: 'curp', headerName: 'CURP', width: 200, hide: true,
-      },
-      { field: 'nombre', headerName: 'Nombre', width: 170 },
-      { field: 'apellido_paterno', headerName: 'Apellido Paterno', width: 190 },
-      { field: 'apellido_materno', headerName: 'Apellido Materno', width: 190 },
-      { field: 'fecha_de_nacimiento', headerName: 'Nacimiento', width: 165 },
-      {
-        field: 'tutor_nombre',
-        headerName: 'Tutor nombre',
-        width: 124,
-        hide: true,
-      },
-      {
-        field: 'tutor_apellido_paterno',
-        headerName: 'Tutor apellido paterno',
-        width: 154,
-        hide: true,
-      },
-      {
-        field: 'tutor_apellido_materno',
-        headerName: 'Tutor apellido materno',
-        width: 160,
-        hide: true,
-      },
-      {
-        field: 'tutor_correo',
-        headerName: 'Tutor correo',
-        width: 154,
-        hide: true,
-      },
-      {
-        field: 'tutor_num_telefono',
-        headerName: 'Tutor telefono',
-        width: 104,
-        hide: true,
-      },
-      { field: 'num_telefono', headerName: 'Telefono', width: 160 },
-      {
-        field: 'pais', headerName: 'Pais', width: 84, hide: true,
-      },
-      {
-        field: 'estado', headerName: 'Estado', width: 84, hide: true,
-      },
-      {
-        field: 'ciudad', headerName: 'Ciudad', width: 84, hide: true,
-      },
-      {
-        field: 'colonia', headerName: 'Colonia', width: 84, hide: true,
-      },
-      {
-        field: 'codigo_postal',
-        headerName: 'codigo postal',
-        width: 84,
-        hide: true,
-      },
-      {
-        field: 'escolaridad',
-        headerName: 'Escolaridad',
-        width: 84,
-        hide: true,
-      },
-      {
-        field: 'ultima_escuela',
-        headerName: 'Ultima escuela',
-        width: 104,
-        hide: true,
-      },
-      {
-        field: 'actions',
-        headerName: 'Acciones',
-        type: 'actions',
-        width: 200,
-        renderCell: (params) => <Actions {...{ params, seleccionarConsola }} />,
-      },
-    ],
-    [studentData],
-  );
-
   const handleSelectChange = async (event) => {
     let array = [];
     let array2 = [];
@@ -359,92 +260,23 @@ export default function Alumnos() {
   return (
     <div>
       <HeaderAlumnos
-      data={studentData}
-      dataPeriodo={dataPeriodo}
-      handleSelectChange={handleSelectChange}
+        data={studentData}
+        dataPeriodo={dataPeriodo}
+        handleSelectChange={handleSelectChange}
       />
-      <Card
-        sx={{
-          width: 1100,
-          position: 'absolute',
-          textAlign: 'left',
-          marginLeft: '65px',
-          marginTop: '120px',
-          bgcolor: 'grey.200',
-          borderRadius: '8px',
-        }}
-      >
-        <CardContent>
-          <Typography
-            gutterBottom
-            variant="h5"
-            component="div"
-            sx={{ textAlign: 'left', fontFamily: 'arial', marginLeft: 1 }}
-          >
-            Filtro
-          </Typography>
-          <TextField
-            style={{
-              paddingBottom: '10px',
-              fontFamily: 'arial',
-              width: 1050,
-              marginLeft: 7,
-            }}
-            label="Ingrese un nombre para buscar"
-            onChange={(e) => {
-              setItems([
-                {
-                  columnField: 'nombre',
-                  operatorValue: 'contains',
-                  value: e.target.value,
-                },
-              ]);
-            }}
-          />
-        </CardContent>
-      </Card>
-      <Box
-        sx={{
-          width: '1130px',
-          padding: '15px',
-          height: '450px',
-          position: 'absolute',
-          marginLeft: '50px',
-          marginTop: '270px',
-        }}
-      >
-        <DataGrid
-          columns={columns}
-          rows={studentData}
-          getRowId={(row) => row._id}
-          rowsPerPageOptions={[5, 10]}
-          pageSize={pageSize}
-          onPageSizeChange={(newPageSize) => SetPageSize(newPageSize)}
-          getRowSpacing={(params) => ({
-            top: params.isFirstVisible ? 0 : 5,
-            bottom: params.isLastVisible ? 0 : 5,
-          })}
-          sx={{
-            [`& .${gridClasses.row}`]: {
-              bgcolor: (theme) => (theme.palette.mode === 'light' ? grey[200] : grey[900]),
-              fontFamily: 'arial',
-            },
-          }}
-          disableSelectionOnClick
-          filterModel={{
-            items,
-          }}
-        />
-        <Modal open={openModal} onClose={abrirCerrarModal}>
-          <MasInformacionAlumno consolaSeleccionada={alumnoSeleccionado} />
-        </Modal>
-        <Modal open={openModal} onClose={abrirCerrarModal}>
-          {bodyEditar}
-        </Modal>
-        <Modal open={openModal} onClose={abrirCerrarModal}>
-          {bodyEliminar}
-        </Modal>
-      </Box>
+      <BodyAlumnos
+        data={studentData}
+        seleccionarConsola={seleccionarAlumno}
+      />
+      <Modal open={openModal} onClose={abrirCerrarModal}>
+        <MasInformacionAlumno consolaSeleccionada={alumnoSeleccionado} />
+      </Modal>
+      <Modal open={openModal} onClose={abrirCerrarModal}>
+        {bodyEditar}
+      </Modal>
+      <Modal open={openModal} onClose={abrirCerrarModal}>
+        {bodyEliminar}
+      </Modal>
     </div>
   );
 }
